Version 12
The partitioning of fluxes relies heavily on estimation of the ustar
threshold.  Up to v11 of DINGO this used the Reichstein approach for
determining ustar threshold.  This is still the most widely used and is
employed in the online Eddyflux processing software.   It is much the
gold standard.
Now V12 is a major step forward.  It employs the Barr et al. method to
calculate ustar.  Ian McHugh has spent a lot of time coding this method
into python and getting this right and testing it (thanks Ian).  Now v12
allows you to calculate Re using Reichstein, Barr, Lasslop or manual
ustar threshold and by default opts to the Barr et al. method.  This
gives a distribution of ustar thresholds (PDF).  This provides a known
uncertainty to the ustar threshold used which is useful for quantifying
errors.  In addition, later on this uncertainty can be propagated
through the fluxes to quantify the impact of the uncertainties in the
ustar threshold on the fluxes 
Many other minor fixes and improvements